Updated ruletest_jsons directory to better mirror rulesets directory, correctly organizing files by ruletest. Updated the excel to JSON scripts to reference this new structure and organized them into a new 'excel_to_test_json_utilities.py' script
