+++ rct229/ruletest_engine/ruletest_jsons/ashrae9012019/user_baseline_rmr.json
@@ -1,0 +1,52 @@
+{
+	"id": "Lighting Example",
+	"ruleset_model_type": "USER"
+	"buildings": [
+		{
+			"id": "Building Name 1",
+			"building_open_schedule": "Required Building Schedule 1",
+			"building_segments": [
+				{
+					"id": "Building Segment 1",
+					"lighting_building_area_type": "HOTEL_MOTEL",
+					"zones": [
+						{
+							"id": "Thermal Zone 1",
+							"volume": 311.48531251199995,
+							"thermostat_cooling_setpoint_schedule": "Required Cooling Schedule 1",
+							"thermostat_heating_setpoint_schedule": "Required Heating Schedule 1",
+							"spaces": [
+								{
+									"id": "Space 1",
+									"lighting_space_type": "GUEST_ROOM",
+									"floor_area": 9.290303999999999,
+									"occupant_multiplier_schedule": "Required Occupant Schedule 1",
+									"interior_lighting": [
+										{
+											"id": "Guest Room Lighting",
+											"power_per_area": 6.5,
+											"lighting_multiplier_schedule": "Required Lighting Multiplier Schedule 1"
+										}
+									]
+								},
+								{
+									"id": "Space 2",
+									"lighting_space_type": "LOBBY_HOTEL",
+									"floor_area": 92.90303999999999,
+									"occupant_multiplier_schedule": "Required Occupant Schedule 1",
+									"interior_lighting": [
+										{
+											"id": "Hotel Lobby Lighting",
+											"power_per_area": 13.1,
+											"lighting_multiplier_schedule": "Required Lighting Multiplier Schedule 1"
+										}
+									]
+								}
+							]
+						}
+					]
+				}
+			]
+		}
+	]
+}
